remove client and dont delete integration tests
