+++ nbs-manage-your-appointments.sln
@@ -23,12 +23,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Core.UnitT
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Persistance.UnitTests", "tests\Nhs.Appointments.Persistance.UnitTests\Nhs.Appointments.Persistance.UnitTests.csproj", "{CBA2FC86-E052-4C1B-A8DD-BB8C07962178}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.ApiClient", "src\api\Nhs.Appointments.ApiClient\Nhs.Appointments.ApiClient.csproj", "{5248046F-8009-4A04-8FFC-B848E704A3A3}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Api.Client.UnitTests", "tests\Nhs.Appointments.Api.Client.UnitTests\Nhs.Appointments.Api.Client.UnitTests.csproj", "{F9E0C737-EC44-48D8-B656-EBCC8F09C033}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Api.Client.Integration", "tests\Nhs.Appointments.Api.Client.Integration\Nhs.Appointments.Api.Client.Integration.csproj", "{DFA538A6-9D01-417F-A0B3-D91102B9FA22}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "local-tools", "local-tools", "{22B97405-2671-42B1-BE08-E1D0E18A2F1F}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Api.Integration", "tests\Nhs.Appointments.Api.Integration\Nhs.Appointments.Api.Integration.csproj", "{6595832E-26F8-4463-A738-1874F4F2DAD9}"
@@ -100,18 +94,6 @@ Global
 		{CBA2FC86-E052-4C1B-A8DD-BB8C07962178}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CBA2FC86-E052-4C1B-A8DD-BB8C07962178}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CBA2FC86-E052-4C1B-A8DD-BB8C07962178}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5248046F-8009-4A04-8FFC-B848E704A3A3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5248046F-8009-4A04-8FFC-B848E704A3A3}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5248046F-8009-4A04-8FFC-B848E704A3A3}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5248046F-8009-4A04-8FFC-B848E704A3A3}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F9E0C737-EC44-48D8-B656-EBCC8F09C033}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F9E0C737-EC44-48D8-B656-EBCC8F09C033}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F9E0C737-EC44-48D8-B656-EBCC8F09C033}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F9E0C737-EC44-48D8-B656-EBCC8F09C033}.Release|Any CPU.Build.0 = Release|Any CPU
-		{DFA538A6-9D01-417F-A0B3-D91102B9FA22}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{DFA538A6-9D01-417F-A0B3-D91102B9FA22}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{DFA538A6-9D01-417F-A0B3-D91102B9FA22}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{DFA538A6-9D01-417F-A0B3-D91102B9FA22}.Release|Any CPU.Build.0 = Release|Any CPU
 		{6595832E-26F8-4463-A738-1874F4F2DAD9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{6595832E-26F8-4463-A738-1874F4F2DAD9}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{6595832E-26F8-4463-A738-1874F4F2DAD9}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -181,9 +163,6 @@ Global
 		{1F4B6550-CA50-47EF-9F28-E9E48680E85E} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
 		{7719F088-61B9-458D-AEE5-6385DE4E91CA} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
 		{CBA2FC86-E052-4C1B-A8DD-BB8C07962178} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
-		{5248046F-8009-4A04-8FFC-B848E704A3A3} = {6F9747AA-DE8B-43B4-B77F-4DDD959BB94C}
-		{F9E0C737-EC44-48D8-B656-EBCC8F09C033} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
-		{DFA538A6-9D01-417F-A0B3-D91102B9FA22} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
 		{6595832E-26F8-4463-A738-1874F4F2DAD9} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
 		{41309612-F5A4-41CF-9391-008EFAD96E80} = {22B97405-2671-42B1-BE08-E1D0E18A2F1F}
 		{00BF2F24-0FB3-420B-BE1C-830919F6EA3E} = {22B97405-2671-42B1-BE08-E1D0E18A2F1F}
