--- src/api/Nhs.Appointments.ApiClient/Auth/IRequestSigner.cs
@@ -1,7 +1,0 @@
-﻿namespace Nhs.Appointments.ApiClient.Auth
-{
-    public interface IRequestSigner
-    {
-        Task SignRequestAsync(HttpRequestMessage request);
-    }
-}
