--- src/api/Nhs.Appointments.ApiClient/INhsAppointmentsApi.cs
@@ -1,9 +1,0 @@
-﻿namespace Nhs.Appointments.ApiClient
-{
-    public interface INhsAppointmentsApi
-    {
-        IBookingsApiClient Bookings { get; }
-        ISitesApiClient Sites { get; }
-        ITemplatesApiClient Templates { get; }
-    }
-}
