--- src/api/Nhs.Appointments.ApiClient/Impl/UnexpectedResponseException.cs
@@ -1,13 +1,0 @@
-﻿using System.Net;
-
-namespace Nhs.Appointments.ApiClient.Impl
-{
-    public class UnexpectedResponseException : Exception 
-    {
-        public UnexpectedResponseException(HttpStatusCode received, HttpStatusCode[] expected)
-            : base($"Recieved HTTP status code {(int)received} {received}. Expected: {string.Join(',', expected)}")
-        {
-
-        }
-    }
-}
