--- src/api/Nhs.Appointments.ApiClient/Models/ContactItem.cs
@@ -1,12 +1,0 @@
-﻿
-using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models;
-
-public record ContactItem(
-[property: JsonPropertyName("type")]
-string Type,
-[property: JsonPropertyName("value")]
-string Value
-)
-{ }
