--- src/api/Nhs.Appointments.ApiClient/Models/Converters/ShortTimeOnlyJsonConverter.cs
@@ -1,22 +1,0 @@
-﻿
-using System.Globalization;
-using System.Text.Json;
-using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models.Converters
-{
-    public class ShortTimeOnlyJsonConverter : JsonConverter<TimeOnly>
-    {
-        private const string TimeFormat = "HH:mm";
-
-        public override TimeOnly Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
-        {
-            return TimeOnly.ParseExact(reader.GetString(), TimeFormat, CultureInfo.InvariantCulture);
-        }
-
-        public override void Write(Utf8JsonWriter writer, TimeOnly value, JsonSerializerOptions options)
-        {
-            writer.WriteStringValue(value.ToString(TimeFormat, CultureInfo.InvariantCulture));
-        }
-    }
-}
