--- src/api/Nhs.Appointments.ApiClient/Models/GetTemplateAssignmentsResponse.cs
@@ -1,6 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    public record GetTemplateAssignmentsResponse([property: JsonPropertyName("assignments")] TemplateAssignment[] Assignments);
-}
