--- src/api/Nhs.Appointments.ApiClient/Models/GetTemplateResponse.cs
@@ -1,10 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    public record GetTemplateResponse
-    {
-        [property: JsonPropertyName("templates")]
-        public WeekTemplate[] Templates { get; set; }
-    }
-}
