--- src/api/Nhs.Appointments.ApiClient/Models/MakeBookingResponse.cs
@@ -1,6 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    public record MakeBookingResponse([property: JsonPropertyName("bookingReference")]string BookingReference);
-}
