--- src/api/Nhs.Appointments.ApiClient/Models/ModelException.cs
@@ -1,10 +1,0 @@
-﻿namespace Nhs.Appointments.ApiClient.Models
-{
-    public class ModelException : Exception
-    {
-        public ModelException(string json, System.Text.Json.JsonException inner) : base($"Json could not be parsed: {json}", inner)
-        {
-
-        }
-    }
-}
