--- src/api/Nhs.Appointments.ApiClient/Models/QueryAvailabilityResponse.cs
@@ -1,4 +1,0 @@
-﻿namespace Nhs.Appointments.ApiClient.Models
-{
-    public class QueryAvailabilityResponse : List<QueryAvailabilityResponseItem> { }
-}
