--- src/api/Nhs.Appointments.ApiClient/Models/QueryType.cs
@@ -1,19 +1,0 @@
-﻿
-using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    [JsonConverter(typeof(JsonStringEnumConverter))]
-    public enum QueryType
-    {
-        NotSet,
-
-        Days,
-
-        Hours,
-
-        Slots
-    }
-
-    
-}
