--- src/api/Nhs.Appointments.ApiClient/Models/ServiceConfiguration.cs
@@ -1,10 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    public record ServiceConfiguration(
-    [property: JsonPropertyName("code")] string Code,
-    [property: JsonPropertyName("displayName")] string DisplayName,
-    [property: JsonPropertyName("duration")] int Duration,
-    [property: JsonPropertyName("enabled")] bool Enabled);
-}
