--- src/api/Nhs.Appointments.ApiClient/Models/SetTemplateAssignmentRequest.cs
@@ -1,13 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    public record SetTemplateAssignmentRequest
-    {
-        [property: JsonPropertyName("site")]
-        public string Site { get; set; }
-
-        [property: JsonPropertyName("assignments")]
-        public TemplateAssignment[] Assignments { get; set; }
-    }
-}
