--- src/api/Nhs.Appointments.ApiClient/Models/TemplateAssignment.cs
@@ -1,10 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Nhs.Appointments.ApiClient.Models
-{
-    public record TemplateAssignment(
-        [property: JsonPropertyName("from")] string From,
-        [property:JsonPropertyName("until")] string Until,
-        [property:JsonPropertyName("templateId")] string TemplateId);
-
-}
